Move compressors to jvm
