+++ core/src/main/scalajs/sttp/client4/compression/CompressorExtensions.scala
@@ -1,0 +1,5 @@
+package sttp.client4.compression
+
+trait CompressorExtensions {
+  def default[R]: List[Compressor[R]] = Nil
+}
