+++ core/src/main/scalajvm/sttp/client4/compression/GZIPCompressingInputStream.scala
